examenes_backend/requirements.txt
fastapi
uvicorn
pydantic
google-generativeai
pymongo
pandas
python-multipart
pdf2image
Pillow
PyPDF2
reportlab
python-dotenv
requests
pytest
pytest-asyncio
httpx
pytest-mock
pytest-cov
fpdf2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4


